+++ gegelatilib/include/tpg/tpgVertex.h
@@ -57,7 +57,7 @@ namespace TPG {
         const std::list<TPGEdge*>& getIncomingEdges() const;
 
         /**
-         * \brief Get a const reference to incoming edges of this TPGVertex.
+         * \brief Get a const reference to outgoing edges of this TPGVertex.
          */
         const std::list<TPGEdge*>& getOutgoingEdges() const;